vdev/ui: Move frost code to own file
